+++ src/views/dashboard/workbench/components/TodoList.vue
@@ -1,6 +1,6 @@
 <script lang="tsx">
   import { defineComponent, computed, unref } from 'compatible-vue';
-  import { List, Icon } from 'ant-design-vue';
+  import { List, Icon, Tag, Tooltip } from 'ant-design-vue';
   import { CollapseContainer } from '@/components/container/index';
 
   import { useDesign } from '@/hooks/core/useDesign';
@@ -35,20 +35,23 @@
                       <span class={`${prefixCls}__item-memo`}>{memo}</span>
                     </div>
                     <div slot="description" class={`${prefixCls}__item-desc`}>
-                      提交人：{sbmter}&nbsp;&nbsp;提交时间：{sbmtTime}
+                      提交人：{sbmter}
+                      <br />
+                      提交时间：{sbmtTime}
                     </div>
                   </List.Item.Meta>
                   <a-button onClick={handleAppr.bind(item)} type="link">
-                    待审批
-                    <Icon type="down" />
+                    <Tag color="blue">待审批</Tag>
                   </a-button>
                 </List.Item>
               );
             })}
           </List>
           <div class={`${prefixCls}__all`}>
-            查看全部
-            <Icon type="right" />
+            <Tooltip placement="topRight">
+              <template slot="title">查看更多</template>
+              <Icon type="ellipsis" />
+            </Tooltip>
           </div>
         </CollapseContainer>
       );
@@ -61,28 +64,34 @@
   .@{prefix-cls} {
     // padding: 24px;
     // background: #fff;
+    position: relative;
+
     &__total {
       display: inline-block;
       width: 20px;
       height: 20px;
       font-size: 12px;
       line-height: 20px;
-      color: #000;
+      color: #fff;
       text-align: center;
-      background: rgba(140, 140, 140, 0.1);
+      background: rgba(255, 0, 0, 0.7);
       border-radius: 50%;
-      opacity: 0.65;
     }
 
     &__all {
+      position: absolute;
+      top: 0;
+      right: 10px;
       height: 56px;
+      font-size: 24px;
       line-height: 56px;
       text-align: center;
-      background: #f9f9f9;
-      border-radius: 0 0 2px 2px;
+      cursor: pointer;
     }
 
     &__item {
+      padding: 8px 0;
+
       &-title {
         font-size: 14px;
         font-weight: normal;
@@ -91,14 +100,14 @@
       }
 
       &-memo {
-        font-size: 14px;
+        font-size: 12px;
         font-weight: normal;
         line-height: 22px;
         color: #7c8087;
       }
 
       &-desc {
-        font-size: 14px;
+        font-size: 12px;
         line-height: 22px;
         color: #7c8087;
       }