+++ src/views/examples/table/Demo16.vue
@@ -46,6 +46,8 @@
         }, 0);
         return [
           {
+            _row: '合计',
+            _index: '平均值',
             no: totalNo,
           },
         ];
@@ -53,27 +55,18 @@
       return () => {
         return (
           <div class="p-4 table-demo">
-            <div>
-              <BasicTable
-                rowSelection={{ type: 'checkbox' }}
-                showSummary={true}
-                summaryFunc={handleSummary}
-                scroll={{ x: 2000 }}
-                api={demoListApi}
-                title="表尾行合计示例"
-                columns={columns}
-              ></BasicTable>
-            </div>
+            <BasicTable
+              rowSelection={{ type: 'checkbox' }}
+              showSummary={true}
+              summaryFunc={handleSummary}
+              scroll={{ x: 2000 }}
+              api={demoListApi}
+              title="表尾行合计示例"
+              columns={columns}
+            />
           </div>
         );
       };
     },
   });
 </script>
-<style lang="less" scoped>
-  .table-demo {
-    & > div {
-      background: #fff;
-    }
-  }
-</style>