+++ src/views/sys/login/LoginForm.vue
@@ -107,7 +107,7 @@
         const { openLoginVerify } = appStore.getProjCfg;
         return (
           <div class={prefixCls}>
-            <h1>系统登陆</h1>
+            <h1>系统登录</h1>
             {openLoginVerify && (
               <VerifyModal
                 src={headImg}
@@ -125,7 +125,7 @@
               class="mt-2"
               onClick={handleLogin}
             >
-              登陆
+              登录
             </a-button>
           </div>
         );