+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+## [1.3.2](https://github.com/anncwb/vue-vben-admin/compare/1.3.1...1.3.2) (2020-08-25)
+
+### ♻ Code Refactoring
+
+- **scrollbar:** 使用 ts 重构 scrollbar 组件 ([d1ff672](https://github.com/anncwb/vue-vben-admin/commit/d1ff672))
+- **table:** 修改表格样式,新增 emptyDataIsShowTable 参数,修复已知 bug ([bd3abc7](https://github.com/anncwb/vue-vben-admin/commit/bd3abc7))
+
+### ✨ Features
+
+- **form:** form 新增 baseColProps 和 emptySpan 参数 ([3f442e1](https://github.com/anncwb/vue-vben-admin/commit/3f442e1))
+
+### 🎫 Chores
+
+- **ci:** 升级 antv 到 1.6.5 ([f770e70](https://github.com/anncwb/vue-vben-admin/commit/f770e70))
+
+### 🐛 Bug Fixes
+
+- **form:** 修复表格内表单计算问题 ([1ff8693](https://github.com/anncwb/vue-vben-admin/commit/1ff8693))
+
+## [1.3.1](https://github.com/anncwb/vue-vben-admin/compare/1.3.0...1.3.1) (2020-08-24)
+
+### ⚡ Performance Improvements
+
+- (组件、插件)demo 页面优化 ([50e1ebc](https://github.com/anncwb/vue-vben-admin/commit/50e1ebc))
+- 图表模块样式优化调整 ([3a7f3de](https://github.com/anncwb/vue-vben-admin/commit/3a7f3de))
+
+### 🎫 Chores
+
+- bump 1.3.0 ([8d4b387](https://github.com/anncwb/vue-vben-admin/commit/8d4b387))
+
+### 🐛 Bug Fixes
+
+- **ci:** 修复控制台警告问题 ([6d623df](https://github.com/anncwb/vue-vben-admin/commit/6d623df))
+- **table:** 修复 showIndexColumn 不生效问题 ([a1ad81c](https://github.com/anncwb/vue-vben-admin/commit/a1ad81c))
+- **table:** 修复表格搜索表单第一次加载闪动问题 ([1e91ac6](https://github.com/anncwb/vue-vben-admin/commit/1e91ac6))
+
+### 💄 Styles
+
+- 界面文字修改 ([f668726](https://github.com/anncwb/vue-vben-admin/commit/f668726))
+
 # [1.3.0](https://github.com/anncwb/vue-vben-admin/compare/1.2.0...1.3.0) (2020-08-18)
 
 ### ⚡ Performance Improvements